aimodel tests passing
